+++ lib/feature/navigation/view_model/navigation_view_model.dart
@@ -2,7 +2,6 @@ import 'package:auto_route/auto_route.dart';
 import 'package:core/core.dart';
 import 'package:x_im_v00r01/feature/navigation/view_model/state/navigation_state.dart';
 import 'package:x_im_v00r01/product/cache/model/user_cache_model.dart';
-import 'package:x_im_v00r01/product/navigation/deeplink/app_router.dart';
 import 'package:x_im_v00r01/product/service/interface/authenction_operation.dart';
 import 'package:x_im_v00r01/product/state/base/base_cubit.dart';
 
@@ -35,15 +34,15 @@ final class NavigationViewModel extends BaseCubit<NavigationState> {
     emit(state.copyWith(selectedindex: index));
 
     if (index == 0) {
-      router.replaceAll([const HomenewRoute()]);
+      router.replaceNamed('story/1');
     } else if (index == 1) {
-      router.replaceAll([const DiscoverRoute()]);
+      router.replaceNamed('discover');
     } else if (index == 2) {
-      router.replaceAll([const OnboardingsRoute()]);
+      router.replaceNamed('onboardings');
     } else if (index == 3) {
-      router.replaceAll([const FavoritesRoute()]);
+      router.replaceNamed('favorites');
     } else if (index == 4) {
-      router.replaceAll([const SettingsRoute()]);
+      router.replaceNamed('settings');
     }
   }
 }
