+++ lib/feature/storyy/view/storyy_view.dart
@@ -21,7 +21,7 @@ class _StoryyViewState extends BaseState<StoryyView> with StoryViewMixin {
         leading: IconButton(
           icon: const Icon(Icons.arrow_back, color: Colors.black),
           onPressed: () {
-            context.router.popAndPush(const HomenewRoute());
+            context.router.popAndPush( HomenewRoute());
           }, // Geri butonu işlevi
         ),
         actions: [
