+++ lib/product/navigation/deeplink/app_router.gr.dart
@@ -114,10 +114,18 @@ class HomeRoute extends PageRouteInfo<void> {
 
 /// generated route for
 /// [HomenewView]
-class HomenewRoute extends PageRouteInfo<void> {
-  const HomenewRoute({List<PageRouteInfo>? children})
-      : super(
+class HomenewRoute extends PageRouteInfo<HomenewRouteArgs> {
+  HomenewRoute({
+    Key? key,
+    dynamic storyId,
+    List<PageRouteInfo>? children,
+  }) : super(
           HomenewRoute.name,
+          args: HomenewRouteArgs(
+            key: key,
+            storyId: storyId,
+          ),
+          rawPathParams: {'storyId': storyId},
           initialChildren: children,
         );
 
@@ -126,9 +134,31 @@ class HomenewRoute extends PageRouteInfo<void> {
   static PageInfo page = PageInfo(
     name,
     builder: (data) {
-      return const HomenewView();
+      final pathParams = data.inheritedPathParams;
+      final args = data.argsAs<HomenewRouteArgs>(
+          orElse: () => HomenewRouteArgs(storyId: pathParams.get('storyId')));
+      return HomenewView(
+        key: args.key,
+        storyId: args.storyId,
+      );
     },
   );
+}
+
+class HomenewRouteArgs {
+  const HomenewRouteArgs({
+    this.key,
+    this.storyId,
+  });
+
+  final Key? key;
+
+  final dynamic storyId;
+
+  @override
+  String toString() {
+    return 'HomenewRouteArgs{key: $key, storyId: $storyId}';
+  }
 }
 
 /// generated route for
